requirements.txt
streamlit
numpy
tensorflow
pillow
streamlit-webrtc
av
mlflow
protobuf==3.20.3
mediapipe
tqdm
mediapipe